+++ include/ccf/historical_queries_interface.h
@@ -59,10 +59,10 @@ namespace ccf::historical
    * is asynchronously retrieved from the ledger and then validated. If the same
    * handle is used for a new range, the state for the old range will be
    * discarded. State is also discarded after the handle's expiry duration, or
-   * when drop_request is called for a given handle. The management of requests
-   * (how many unique handles are concurrently active, how they are correlated
-   * across HTTP requests, how the active quota is divided between callers) is
-   * left to the calling system.
+   * when drop_cached_states is called for a given handle. The management of
+   * requests (how many unique handles are concurrently active, how they are
+   * correlated across HTTP requests, how the active quota is divided between
+   * callers) is left to the calling system.
    */
   class AbstractStateCache
   {
@@ -115,7 +115,7 @@ namespace ccf::historical
      * to completely fetch and validate. The call should be repeated later with
      * the same arguments to retrieve the requested entries. This state is kept
      * until it is deleted for one of the following reasons:
-     *  - A call to @c drop_request
+     *  - A call to @c drop_cached_states
      *  - @c seconds_until_expiry seconds elapse without calling this function
      *  - This handle is used to request a different seqno or range
      *
@@ -155,8 +155,8 @@ namespace ccf::historical
     /** Drop state for the given handle.
      *
      * May be used to free up space once a historical query has been resolved,
-     * more aggressively than waiting for the requests to expire.
+     * more aggressively than waiting for the states to expire.
      */
-    virtual bool drop_request(RequestHandle handle) = 0;
+    virtual bool drop_cached_states(RequestHandle handle) = 0;
   };
 }