+++ samples/apps/logging/logging.cpp
@@ -1054,7 +1054,7 @@ namespace loggingapp
         // ALSO: Assume this response makes it all the way to the client, and
         // they're finished with it, so we can drop the retrieved state. In a
         // real app this may be driven by a separate client request or an LRU
-        historical_cache.drop_request(handle);
+        historical_cache.drop_cached_states(handle);
       };
       make_endpoint(
         get_historical_range_path,