+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Renamed
+
+- `ccf::historical::AbstractStateCache::drop_requests()` renamed to `drop_cached_states()`
+
 ## [2.0.0-dev6]
 
 ### Added